HW9 — Kafka/requirements.txt
kafka_python==2.2.15
